disproject-custom-suffixes: Update type
When custom suffixes were revised to accept Transient syntax, the custom
suffixes variable type was not updated to handle it.

Since we're delegating syntax to be handled by Transient anyways, the effort
required to maintain correct types for something that'll be checked elsewhere
anyways seems unnecessary, so set the type to 'sexp.