Add dispatch menu for finding "special files" in projects
Facilities for extending the menu more easily will be added later.

Remove comment about `disproject--custom-suffixes-prefix` being a hack.  It'll
remain private (for the foreseeable future), but I think ensuring values in the
scope like this is actually the most reasonable place to do it.  It's possibly
the earliest point where we can set a prefix's scope (doing it in the prefix
command's body means it's possible to unintentionally step around it if setup
occurs in another function), while doing it any later means we could run into
some kind of race condition where e.g. the scope is ensured for some
descriptions but not for others.